moved Driver class into its own file
